[Feature-12-resource-availability] Refactored so there is ResourceAvailabilityService and new test file for it.
