+++ src/main/java/jewellery/inventory/controller/ResourceAvailabilityController.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 import jewellery.inventory.dto.request.resource.ResourceInUserRequestDto;
 import jewellery.inventory.dto.response.UserResponseDto;
 import jewellery.inventory.dto.response.resource.ResourceInUserResponseDto;
-import jewellery.inventory.service.UserService;
+import jewellery.inventory.service.ResourceAvailabilityService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -24,37 +24,37 @@ import org.springframework.web.bind.annotation.RestController;
 @CrossOrigin(origins = "${cors.origins}")
 @RequiredArgsConstructor
 public class ResourceAvailabilityController {
-  private final UserService userService;
+  private final ResourceAvailabilityService resourceAvailabilityService;
 
   @ResponseStatus(HttpStatus.CREATED)
   @PostMapping
   public UserResponseDto addResourceToUser(
       @RequestBody @Valid ResourceInUserRequestDto resourceUserDto) {
-    return userService.addResourceToUser(resourceUserDto);
+    return resourceAvailabilityService.addResourceToUser(resourceUserDto);
   }
 
   @ResponseStatus(HttpStatus.OK)
   @GetMapping("/{userId}/{resourceId}")
   public double getUserResourceQuantity(@PathVariable UUID userId, @PathVariable UUID resourceId) {
-    return userService.getUserResourceQuantity(userId, resourceId);
+    return resourceAvailabilityService.getUserResourceQuantity(userId, resourceId);
   }
 
   @ResponseStatus(HttpStatus.OK)
   @GetMapping("/{userId}")
   public List<ResourceInUserResponseDto> getAllResourcesFromUser(@PathVariable UUID userId) {
-    return userService.getAllResourcesFromUser(userId);
+    return resourceAvailabilityService.getAllResourcesFromUser(userId);
   }
 
   @ResponseStatus(HttpStatus.NO_CONTENT)
   @DeleteMapping("/{userId}/{resourceId}")
   public void removeResourceFromUser(@PathVariable UUID userId, @PathVariable UUID resourceId) {
-    userService.removeResourceFromUser(userId, resourceId);
+    resourceAvailabilityService.removeResourceFromUser(userId, resourceId);
   }
 
   @ResponseStatus(HttpStatus.NO_CONTENT)
   @DeleteMapping("/{userId}/{resourceId}/{quantity}")
   public void removeQuantityFromUserResource(
       @PathVariable UUID userId, @PathVariable UUID resourceId, @PathVariable int quantity) {
-    userService.removeQuantityFromResource(userId, resourceId, quantity);
+    resourceAvailabilityService.removeQuantityFromResource(userId, resourceId, quantity);
   }
 }
